+++ modules/@apostrophecms/admin-bar/ui/apos/components/TheAposContextTitle.vue
@@ -102,9 +102,24 @@ export default {
           modifiers: (this.draftMode === 'published') ? [ 'disabled', 'selected' ] : null
         }
       ];
+    },
+    canTogglePublishDraftMode() {
+      return !this.isUnpublished && !this.hasCustomUi;
     }
   },
+  mounted() {
+    apos.bus.$on('command-menu-admin-bar-toggle-publish-draft', this.togglePublishDraftMode);
+  },
+  destroyed() {
+    apos.bus.$off('command-menu-admin-bar-toggle-publish-draft', this.togglePublishDraftMode);
+  },
   methods: {
+    togglePublishDraftMode() {
+      if (this.canTogglePublishDraftMode) {
+        const mode = this.draftMode === 'draft' ? 'published' : 'draft';
+        this.switchDraftMode(mode);
+      }
+    },
     switchDraftMode(mode) {
       this.$emit('switch-draft-mode', mode);
     }