+++ modules/@apostrophecms/doc-type/ui/apos/components/AposDocContextMenu.vue
@@ -320,11 +320,18 @@ export default {
   },
   mounted() {
     apos.bus.$on('content-changed', this.onContentChanged);
+    apos.bus.$on('command-menu-admin-bar-discard-draft', this.customDiscardDraft);
   },
   destroyed() {
     apos.bus.$off('content-changed', this.onContentChanged);
+    apos.bus.$off('command-menu-admin-bar-discard-draft', this.customDiscardDraft);
   },
   methods: {
+    customDiscardDraft() {
+      if (this.showDiscardDraft && this.canDiscardDraft) {
+        this.menuHandler('discardDraft');
+      }
+    },
     async onContentChanged(e) {
       if (e.doc && (e.doc._id === this.context._id)) {
         this.context = e.doc;