+++ modules/@apostrophecms/page/index.js
@@ -48,6 +48,43 @@ module.exports = {
       }
     }
   },
+  commands(self) {
+    return {
+      add: {
+        [`${self.__meta.name}:create-new`]: {
+          type: 'item',
+          label: 'apostrophe:commandMenuCreateNew',
+          action: {
+            type: 'command-menu-manager-create-new'
+          },
+          permission: {
+            action: 'edit',
+            type: self.__meta.name
+          },
+          shortcut: 'C'
+        },
+        [`${self.__meta.name}:exit-manager`]: {
+          type: 'item',
+          label: 'apostrophe:commandMenuExitManager',
+          action: {
+            type: 'command-menu-manager-close'
+          },
+          shortcut: 'Q'
+        }
+      },
+      modal: {
+        [`${self.__meta.name}:manager`]: {
+          '@apostrophecms/command-menu:manager': {
+            label: '',
+            commands: [
+              `${self.__meta.name}:create-new`,
+              `${self.__meta.name}:exit-manager`
+            ]
+          }
+        }
+      }
+    };
+  },
   async init(self) {
     const { enableCacheOnDemand = true } = self.apos
       .modules['@apostrophecms/express'].options;